Update tooltip documentation to include new formatter
The documentation for tooltip point text has been updated to include the `ToolTipLabelFormatter` property in addition to the existing `YToolTipLabelFormatter` and `XToolTipLabelFormatter` properties. This change clarifies that users can now use any of these three properties to define the text displayed in tooltips for a given point, providing more flexibility in customizing tooltip content.