Add SVG path handling to GeometryVisual class
Added a new property `Svg` of type `string?` to the `GeometryVisual` class, enabling it to get or set an SVG path. Implemented logic to update the SVG path of the geometry if the `Svg` property is set and the geometry implements the `IVariableSvgPath<TDrawingContext>` interface. This enhances the class's capability to work with SVG data.