Fix AxisTesting Min/Max limits for x2 and y2 axes
Corrected the MinLimit and MaxLimit values in the AxisTesting class
for the x2 and y2 axes. Previously, MaxLimit was set to 0 and
MinLimit was set to 10, which was incorrect. The values have been
swapped so that MinLimit is now 0 and MaxLimit is now 10, while
keeping the IsInverted property set to true.